[Feature][Plugin] Package plugins by category and isolate classloaders
