+++ scaleph-common/src/main/java/cn/sliew/scaleph/common/util/TemporaryClassLoaderContext.java
@@ -1,0 +1,41 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+
+package cn.sliew.scaleph.common.util;
+
+public class TemporaryClassLoaderContext implements AutoCloseable {
+
+    private final ClassLoader originalClassLoader;
+    private final Thread originalThread;
+
+    private TemporaryClassLoaderContext(Thread thread, ClassLoader classLoader) {
+        this.originalThread = thread;
+        this.originalClassLoader = thread.getContextClassLoader();
+        this.originalThread.setContextClassLoader(classLoader);
+    }
+
+    public static TemporaryClassLoaderContext of(ClassLoader classLoader) {
+        return new TemporaryClassLoaderContext(Thread.currentThread(), classLoader);
+    }
+
+    @Override
+    public void close() throws Exception {
+        this.originalThread.setContextClassLoader(this.originalClassLoader);
+    }
+}
