+++ scaleph-dist/src/assembly/scaleph-dist.xml
@@ -77,6 +77,91 @@
             <outputDirectory>.</outputDirectory>
             <fileMode>0644</fileMode>
         </fileSet>
+        <!-- Datasource plugins -->
+        <fileSet>
+            <directory>../scaleph-plugins/scaleph-plugin-datasource/scaleph-plugin-datasource-clickhouse/target
+            </directory>
+            <includes>
+                <include>*.jar</include>
+                <include>dependencies/*.jar</include>
+            </includes>
+            <outputDirectory>plugins/datasource/clickhouse</outputDirectory>
+            <fileMode>0644</fileMode>
+        </fileSet>
+        <fileSet>
+            <directory>../scaleph-plugins/scaleph-plugin-datasource/scaleph-plugin-datasource-doris/target</directory>
+            <includes>
+                <include>*.jar</include>
+                <include>dependencies/*.jar</include>
+            </includes>
+            <outputDirectory>plugins/datasource/doris</outputDirectory>
+            <fileMode>0644</fileMode>
+        </fileSet>
+        <fileSet>
+            <directory>../scaleph-plugins/scaleph-plugin-datasource/scaleph-plugin-datasource-druid/target</directory>
+            <includes>
+                <include>*.jar</include>
+                <include>dependencies/*.jar</include>
+            </includes>
+            <outputDirectory>plugins/datasource/druid</outputDirectory>
+            <fileMode>0644</fileMode>
+        </fileSet>
+        <fileSet>
+            <directory>../scaleph-plugins/scaleph-plugin-datasource/scaleph-plugin-datasource-elasticsearch/target
+            </directory>
+            <includes>
+                <include>*.jar</include>
+                <include>dependencies/*.jar</include>
+            </includes>
+            <outputDirectory>plugins/datasource/elasticsearch</outputDirectory>
+            <fileMode>0644</fileMode>
+        </fileSet>
+        <fileSet>
+            <directory>../scaleph-plugins/scaleph-plugin-datasource/scaleph-plugin-datasource-greenplum/target
+            </directory>
+            <includes>
+                <include>*.jar</include>
+                <include>dependencies/*.jar</include>
+            </includes>
+            <outputDirectory>plugins/datasource/greenplum</outputDirectory>
+            <fileMode>0644</fileMode>
+        </fileSet>
+        <fileSet>
+            <directory>../scaleph-plugins/scaleph-plugin-datasource/scaleph-plugin-datasource-kafka/target</directory>
+            <includes>
+                <include>*.jar</include>
+                <include>dependencies/*.jar</include>
+            </includes>
+            <outputDirectory>plugins/datasource/kafka</outputDirectory>
+            <fileMode>0644</fileMode>
+        </fileSet>
+        <fileSet>
+            <directory>../scaleph-plugins/scaleph-plugin-datasource/scaleph-plugin-datasource-mysql/target</directory>
+            <includes>
+                <include>*.jar</include>
+                <include>dependencies/*.jar</include>
+            </includes>
+            <outputDirectory>plugins/datasource/mysql</outputDirectory>
+            <fileMode>0644</fileMode>
+        </fileSet>
+        <fileSet>
+            <directory>../scaleph-plugins/scaleph-plugin-datasource/scaleph-plugin-datasource-oracle/target</directory>
+            <includes>
+                <include>*.jar</include>
+                <include>dependencies/*.jar</include>
+            </includes>
+            <outputDirectory>plugins/datasource/oracle</outputDirectory>
+            <fileMode>0644</fileMode>
+        </fileSet>
+        <fileSet>
+            <directory>../scaleph-plugins/scaleph-plugin-datasource/scaleph-plugin-datasource-postgre/target</directory>
+            <includes>
+                <include>*.jar</include>
+                <include>dependencies/*.jar</include>
+            </includes>
+            <outputDirectory>plugins/datasource/postgre</outputDirectory>
+            <fileMode>0644</fileMode>
+        </fileSet>
     </fileSets>
 
     <dependencySets>
