+++ scaleph-plugins/scaleph-plugin-datasource/scaleph-plugin-datasource-clickhouse/src/main/java/cn/sliew/scaleph/plugin/datasource/clickhouse/ClickhouseProperties.java
