+++ scaleph-plugins/scaleph-plugin-datasource/scaleph-plugin-datasource-doris/src/main/java/cn/sliew/scaleph/plugin/datasource/doris/DorisProperties.java
