+++ scaleph-plugins/scaleph-plugin-datasource/scaleph-plugin-datasource-elasticsearch/src/main/java/cn/sliew/scaleph/plugin/datasource/elasticsearch/ElasticsearchProperties.java
