+++ scaleph-plugins/scaleph-plugin-datasource/scaleph-plugin-datasource-jdbc/src/main/java/cn/sliew/scaleph/plugin/datasource/jdbc/JdbcPoolProperties.java
