+++ scaleph-plugins/scaleph-plugin-datasource/scaleph-plugin-datasource-kafka/src/main/java/cn/sliew/scaleph/plugin/datasource/kafka/KafkaProperties.java
