--- scaleph-plugins/scaleph-plugin-datasource/src/main/resources/plugin.properties
@@ -1,4 +1,0 @@
-name=HIKARICP_JDBC_DATA_SOURCE
-description=Hikaricp Jdbc DataSource
-version=1.0.0
-classname=cn.sliew.scaleph.plugin.datasource.jdbc.JDBCDataSourcePlugin
